modules/feature_pipeline/pyproject.toml
[tool.poetry]
name = "geo-mapping"
version = "0.1.0"
description = "poetry for features pipeline LengoAI"
authors = ["MihaiDavid05 <davidmihai98@yahoo.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
antlr4-python3-runtime = "4.9.3"
anyio = "4.2.0"
async-timeout = "4.0.3"
attrs = "23.1.0"
backoff = "2.2.1"
bokeh = "3.4.1"
branca = "0.7.2"
cachetools = "5.3.2"
certifi = "2024.6.2"
cffi = "1.16.0"
cfgv = "3.4.0"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
click-plugins = "1.1.1"
cligj = "0.7.2"
cloudpickle = "3.0.0"
colorama = "0.4.6"
colorbrewer = "0.2.0"
colour = "0.1.5"
contourpy = "1.2.1"
cryptography = "42.0.7"
cycler = "0.12.1"
decorator = "5.1.1"
defusedxml = "0.7.1"
dill = "0.3.7"
distlib = "0.3.8"
distro = "1.8.0"
docstring-parser = "0.16"
et-xmlfile = "1.1.0"
exceptiongroup = "1.2.0"
filelock = "3.14.0"
fiona = "1.9.6"
folium = "0.16.0"
fonttools = "4.51.0"
frozenlist = "1.4.1"
fsspec = "2024.6.0"
geographiclib = "2.0"
geopandas = "0.14.4"
geopy = "2.4.1"
google-api-core = "2.15.0"
google-api-python-client = "2.125.0"
google-auth = "2.25.2"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.0"
google-cloud-aiplatform = "1.48.0"
google-cloud-appengine-logging = "1.4.3"
google-cloud-audit-log = "0.2.5"
google-cloud-bigquery = "3.14.1"
google-cloud-core = "2.4.1"
google-cloud-logging = "3.10.0"
google-cloud-resource-manager = "1.11.0"
google-cloud-storage = "2.14.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.0"
googleapis-common-protos = "1.62.0"
grpc-google-iam-v1 = "0.13.0"
grpcio = "1.60.0"
grpcio-status = "1.60.0"
h11 = "0.14.0"
httpcore = "1.0.2"
httplib2 = "0.22.0"
httpx = "0.25.2"
hydra-core = "1.3.2"
identify = "2.5.36"
idna = "3.7"
igraph = "0.11.3"
importlib-metadata = "7.1.0"
importlib-resources = "6.4.0"
jinja2 = "3.1.3"
joblib = "1.4.0"
kiwisolver = "1.4.5"
locket = "1.0.0"
lz4 = "4.3.3"
markupsafe = "2.1.5"
matplotlib = "3.7.5"
msgpack = "1.0.8"
multidict = "6.0.5"
multiprocess = "0.70.15"
mypy = "1.10.0"
mypy-extensions = "1.0.0"
ndg-httpsclient = "0.5.1"
nest-asyncio = "1.5.9"
networkx = "3.1"
nodeenv = "1.9.0"
oauth2client = "4.1.3"
oauthlib = "3.2.2"
omegaconf = "2.3.0"
openai = "1.5.0"
opencv-python = "4.10.0.84"
opencv-python-headless = "4.10.0.84"
openpyxl = "3.1.2"
outcome = "1.3.0.post0"
packaging = "23.2"
pandas = "2.0.3"
partd = "1.4.2"
pillow = "10.1.0"
platformdirs = "4.2.2"
plotly = "5.18.0"
pluscodes = "2022.1.3"
pre-commit = "3.7.1"
proto-plus = "1.23.0"
protobuf = "4.25.1"
psutil = "5.9.8"
py4cytoscape = "1.8.0"
pyarrow = "16.1.0"
pyarrow-hotfix = "0.6"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pydantic = "2.5.2"
pydantic-core = "2.14.5"
pyogrio = "0.8.0"
pyopenssl = "24.1.0"
pyparsing = "3.1.2"
pyproj = "3.6.1"
pyshp = "2.3.1"
pysocks = "1.7.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
rsa = "4.9"
scikit-learn = "1.3.2"
selenium = "^4.23.1"
shapely = "2.0.2"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
supervision = "^0.24.0"
tblib = "3.0.0"
tenacity = "8.2.3"
texttable = "1.7.0"
threadpoolctl = "3.4.0"
tomli = "2.0.1"
toolz = "0.12.1"
tornado = "6.4.1"
tqdm = "4.66.1"
trio = "0.23.1"
trio-websocket = "0.11.1"
typing-extensions = "4.9.0"
tzdata = "2023.3"
uritemplate = "4.1.1"
urllib3 = "2.1.0"
vertexai = "0.0.1"
virtualenv = "20.26.2"
webdriver-manager = "4.0.2"
wsproto = "1.2.0"
xyzservices = "2024.4.0"
yarl = "1.9.4"
zict = "3.0.0"
zipp = "3.18.1"
pykml = "^0.2.0"
haversine = "^2.8.1"
ijson = "^3.3.0"
numba = "^0.60.0"
unidecode = "^1.3.8"
roboflow = "^1.1.49"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
